requirements.txt
# Core dependencies
python-dotenv>=0.19.0
letta>=1.0.0 